+++ rock_paper_scissors/valid_choices.py
@@ -1,4 +1,4 @@
-from moves import Scissors, Paper, Rock
+from rock_paper_scissors.moves import Scissors, Paper, Rock
 
 VALID_CHOICES = {
     'rock': Rock(),
